third_party/WebKit/LayoutTests/bluetooth/descriptor/readValue/read-succeeds.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
const test_desc = "A read request succeeds and returns the descriptor's value.";
const EXPECTED_VALUE = [0, 1, 2];
let descriptor, fake_descriptor;
promise_test(() => getUserDescriptionDescriptor()
  .then(_ => ({descriptor, fake_descriptor} = _))
  .then(() => fake_descriptor.setNextReadResponse(GATT_SUCCESS, EXPECTED_VALUE))
  .then(() => descriptor.readValue())
  .then(value => assert_array_equals(
      Array.from(new Uint8Array(value.buffer)), EXPECTED_VALUE))
  .then(() => navigator.bluetooth.test.allResponsesConsumed())
  .then(consumed => assert_true(consumed)), test_desc);
</script>
